+++ handlers/index.js
@@ -1,34 +1,45 @@
 const handleWelcome = require('./welcome');
 const handleBan = require('./ban');
-const handleUnban = require('./unban');
+const handleUnBan = require('./unBan');
 const handleHelp = require('./help');
 const handleF = require('./f');
 const handleTest = require('./test');
-const handleGifID = require('./gifid');
 const handleKick = require('./kick');
 const handleShrug = require('./shrug');
 const handleReg = require('./reg');
-const handleSetRules = require('./setrules');
+const handleSetRules = require('./setRules');
 const handleRules = require('./rules');
-const handleDurka = require('./durka');
-const handleStickerID = require('./stickerid');
+const handleFun = require('./fun');
+const handleID = require('./id');
+const handleUser = require('./user');
+const handleAdmin = require('./admin');
+const handleLanguage = require('./language');
+const handleGetLink = require('./getLink');
+const handleNoHello = require('./noHello');
+// const handleStickerID = require('./stickerId');
 const handleBomb = require('./bomb');
 const handleDice = require('./dice');
+const handleDarts = require('./darts');
 module.exports = {
   handleWelcome,
   handleBan,
-  handleUnban,
+  handleUnBan,
   handleHelp,
   handleF,
   handleTest,
-  handleGifID,
   handleKick,
   handleReg,
   handleShrug,
   handleSetRules,
   handleRules,
-  handleStickerID,
-  handleDurka,
+  handleFun,
+  handleID,
+  handleUser,
+  handleAdmin,
+  handleLanguage,
+  handleGetLink,
+  handleNoHello,
   handleBomb,
   handleDice,
+  handleDarts,
 };