+++ handlers/rules.js
@@ -1,7 +1,6 @@
 const {Group} = require('../database');
 module.exports = async (ctx) => {
   const group = await Group.findOne({group_id: ctx.chat.id});
-  const rule = group.rules;
   if (!group) {
     try {
       const answer = ctx.i18n.t('group.notfound');
@@ -15,29 +14,18 @@ module.exports = async (ctx) => {
       );
     }
   } else {
-    if (rule !== '') {
-      try {
-        ctx.reply(rule,
-            {reply_to_message_id: ctx.message.message_id},
-        );
-      } catch (error) {
-        const answer = ctx.i18n.t('error', {error: error});
-        ctx.replyWithMarkdown(answer,
-            {reply_to_message_id: ctx.message.message_id},
-        );
-      }
-    } else {
-      try {
-        const answer = ctx.i18n.t('rule.err');
-        ctx.reply(answer,
-            {reply_to_message_id: ctx.message.message_id},
-        );
-      } catch (error) {
-        const answer = ctx.i18n.t('error', {error: error});
-        ctx.replyWithMarkdown(answer,
-            {reply_to_message_id: ctx.message.message_id},
-        );
-      }
+    try {
+      const rule = group.rules;
+      const isRule = !!rule;
+      const answer = isRule ? rule : ctx.i18n.t('rule.err');
+      ctx.reply(answer,
+          {reply_to_message_id: ctx.message.message_id},
+      );
+    } catch (error) {
+      const answer = ctx.i18n.t('error', {error: error});
+      ctx.replyWithMarkdown(answer,
+          {reply_to_message_id: ctx.message.message_id},
+      );
     }
   };
 };