+++ handlers/test.js
@@ -1,5 +1,5 @@
 const {version} = require('../package.json');
-module.exports = (ctx) => {
+module.exports = async (ctx) => {
   try {
     const answer = ctx.i18n.t('test', {
       user: ctx.from.first_name,
@@ -7,18 +7,17 @@ module.exports = (ctx) => {
       user_id: ctx.from.id,
       version: version,
     });
-    ctx.replyWithMarkdown(answer,
+    const message = await ctx.replyWithMarkdown(answer,
         {reply_to_message_id: ctx.message.message_id},
     );
+    setTimeout(() => {
+      ctx.deleteMessage(message.message_id);
+      ctx.deleteMessage(ctx.message.message_id);
+    }, 5 * 1000);
   } catch (error) {
     const answer = ctx.i18n.t('error', {error: error});
     ctx.replyWithMarkdown(answer,
         {reply_to_message_id: ctx.message.message_id},
     );
   }
-  /* setTimeout(() => {
-    // FIXME: autodeleting
-    ctx.deleteMessage(ctx.message.message_id);
-  }, 5 * 1000);
-  */
 };